refactor: token 관련된 클래스들 패키지 집약
