+++ backend/src/main/java/com/woowacourse/f12/application/auth/token/RefreshTokenRepository.java
@@ -1,4 +1,4 @@
-package com.woowacourse.f12.application.auth;
+package com.woowacourse.f12.application.auth.token;
 
 import java.util.Optional;
