+++ backend/src/main/java/com/woowacourse/f12/presentation/auth/AuthArgumentResolver.java
@@ -1,6 +1,6 @@
 package com.woowacourse.f12.presentation.auth;
 
-import com.woowacourse.f12.application.auth.JwtProvider;
+import com.woowacourse.f12.application.auth.token.JwtProvider;
 import com.woowacourse.f12.exception.unauthorized.TokenInvalidFormatException;
 import java.util.Objects;
 import org.springframework.core.MethodParameter;
