+++ backend/src/test/java/com/woowacourse/f12/application/auth/RefreshTokenTest.java
@@ -2,6 +2,7 @@ package com.woowacourse.f12.application.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.woowacourse.f12.application.auth.token.RefreshToken;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.Test;
